multimodal-vqa-rag/requirements.txt
# ============================================================
# Multimodal VQA RAG Pipeline — requirements.txt
# Python 3.10+  |  Mac (CPU inference only, no training)
# Install: pip install -r requirements.txt
# ============================================================

# ── PyTorch (CPU only, inference only) ───────────────────────
torch==2.3.1
torchvision==0.18.1

# ── HuggingFace (model loading + inference) ──────────────────
transformers==4.45.2
tokenizers==0.20.0
huggingface-hub==0.24.6
safetensors==0.4.5
accelerate==0.34.2

# ── Multimodal Embedding — Nomic Embed Multimodal 3B ─────────
sentence-transformers==3.1.1
einops==0.8.0
timm==1.0.9

# ── Reranker — BGE-reranker-base ─────────────────────────────
FlagEmbedding==1.2.11

# ── Sparse Retrieval — BM25 ──────────────────────────────────
rank-bm25==0.2.2
nltk==3.9.1

# ── Vector Store — ChromaDB ──────────────────────────────────
chromadb==0.5.5

# ── LLM — Ollama (Qwen2.5-VL local) ─────────────────────────
ollama==0.3.3

# ── RAG Orchestration — LlamaIndex ───────────────────────────
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-ollama
# ── Image Processing ─────────────────────────────────────────
Pillow==10.4.0
opencv-python-headless==4.10.0.84

# ── Data Handling ────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2

# ── Evaluation ───────────────────────────────────────────────
scikit-learn==1.5.1

# ── Config + Environment ─────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

# ── CLI ───────────────────────────────────────────────────────
typer==0.12.5
rich==13.8.0

# ── Logging ──────────────────────────────────────────────────
loguru==0.7.2

# ── HTTP ─────────────────────────────────────────────────────
httpx==0.27.2

# ── Testing ──────────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8

# ── UI — Streamlit (Sprint 7) ────────────────────────────────
streamlit==1.37.1

# ── Utilities ────────────────────────────────────────────────
tqdm==4.66.5
tenacity==8.5.0